Optimization, move silenceThreshold to percents (0-100)

+++ subprojects/engine/src/karaed/engine/audio/RangeParams.java
@@ -2,7 +2,7 @@ package karaed.engine.audio;
 
 public interface RangeParams {
 
-    float silenceThreshold(int frame);
+    int silenceThreshold(int frame);
 
     int maxSilenceGap(int frame);
